+++ VehicleTrackingAPI/VTA.Buckets/Models/Vehicle.cs
@@ -12,12 +12,12 @@
 
         public string DeviceId { get; set; }
 
-        public decimal? LastLongtitude { get; set; }
+        public double? LastLongtitude { get; set; }
 
-        public decimal? LastLatitude { get; set; }
+        public double? LastLatitude { get; set; }
 
-        //public decimal Fuel { get; set; } 
+        //public double Fuel { get; set; } 
 
-        //public decimal Speed { get; set; }
+        //public double Speed { get; set; }
     }
 }
